vendor_app/serializers/purchase.py
from rest_framework import serializers
from vendor_app.models import PurchaseOrder

class PurchaseOrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = PurchaseOrder
        fields = '__all__'